% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myScatterplot.R
\name{myScatterplot}
\alias{myScatterplot}
\title{myScatterplot}
\usage{
myScatterplot(x, y, z, stages, pnum)
}
\arguments{
\item{x}{A vector of x-coordinates}

\item{y}{A vector of y-coordinates}

\item{z}{A vector of z-coordinates}

\item{stages}{Number of stages}

\item{pnum}{Number of points}
}
\value{
An htmlwidget
}
\description{
<Add Description>
}
